Chore: Migrate envFromArgs, parameters, and randomClipperPort to TypeScript
Change by Claude Code
