+++ .eslintignore
@@ -1372,6 +1372,7 @@ packages/lib/debug/DebugService.js
 packages/lib/determineBaseAppDirs.js
 packages/lib/dom.js
 packages/lib/downloadController.js
+packages/lib/envFromArgs.js
 packages/lib/errorUtils.js
 packages/lib/errors.js
 packages/lib/eventManager.js
@@ -1472,7 +1473,9 @@ packages/lib/net-utils.js
 packages/lib/ntp.js
 packages/lib/onedrive-api.test.js
 packages/lib/onedrive-api.js
+packages/lib/parameters.js
 packages/lib/path-utils.js
+packages/lib/randomClipperPort.js
 packages/lib/reducer.test.js
 packages/lib/reducer.js
 packages/lib/registry.test.js
